Added NewDecode.v which I forgot to add before.  This is just like Decode.v but uses NewParser.v
In NewDecode.v, you have at the bottom a function called build_parser.
When run with appropriate fuel (say 250), it will build a table-based
parser for instructions and return it as a function of type:

  (list nat) -> option ((list nat) * (list instruction))

where the input is the list of bytes to be parsed, and the output
is the unconsumed list of bytes, and a list of instructions that
get returned by the first matching sequence of bytes.  

When I extract this and run build_parser, it takes about 12 seconds
on my laptop to build the tables.  I haven't yet run the parser that
gets generated to see how it performs.  

Still need to do lots of proofs...