Cleaned up the code in NewParser a bit.  Also, I added a new partial function:
  to_string(a:astgram) : interp(astgram_type a) -> option (list char_p)

and proofs that:

  to_string a v = Some s <-> in_astgram a s v

So, there's our AST pretty printer!  This turned out to be (almost)
trivial.  The only tricky situation is that our treatment of Kleene
star demands that to convert a non-empty list [v1,...,vn] into
a string using the grammar (aStar g), then converting each of
the vi's into string using g must result in non-empty strings.  

But otherwise, it's just a straightforward recursive traversal of
the grammar, as the types ensure there is no non-determinism
(only partiality).