Finished half the proof that the table-driven parser is correct.  That is, I have a proof that if you build the table and initial parser state, and then run parse_token on a list of tokens tks, then you'll get out semantic values vs, such that each v in vs is related to the consumed tokens by the original grammar.
What's missing is the completeness -- i.e., that if
something is in the denotation of the grammar, then
we'll definitely return it when we use the table-
driven parser.

The problem with this direction is twofold:  First,
if the top-level grammar accepts the empty string,
then our parser won't work.  It always consumes at
least one token.  Second, the driver loop for the 
table-driven parser only returns the first parse.  
But I'll work on getting an appropriate statement out.