Added a new file Recognizer.v that provides the code and proofs needed for building the table-based recognizer needed for the RockSalt verifier. This used to be in Parser.v but when I was reworking the definitions in there, I omitted this stuff.  I figured it was probably better to break it out to avoid confusion with the table-based parser.
This is much simpler than the old version, but we might even be able
to make it simpler by basically taking the "DFA" structure built for
the parser, and erasing the semantic actions (and related stuff) to
get out a recognizer.  But instead of doing that, I simply ported the
old definitions and proofs to the new setting.  

This should allow Suman to start making progress in the RockSalt
directory.