Added support for a "vector" module that allows us to get constant-time access in extracted ML code.
Provided a vector-based DFA (as opposed to list-based) in 
the Parser module.  I still need to add code so that the x86 
parser uses the vector-based version and test it out.  

I'm leaving all of the other code untouched so we can 
easily compare to see if this is a win.