Changed Vectors to use unsafe_get (avoid bounds check).
Changed vparse_token so that it doesn't run the fixup-function
on each token.  Rather, it expects the caller to run it when 
the caller gets a non-empty list of output values.  

Slightly better flat_map that avoids running append when we
have a singleton list.  

Finished the proofs for vparse_token.

I haven't bothered to prove vparse_tokens since we don't use
that.